# Reduce allocations in `pgn.jl` by preallocating `IOBuffer` in `PGNReader`