@romstad 2022-03-16
No need to apologize for the influx of PRs, they are very welcome! These are all great improvements.

It is I who should apologize for reacting so slowly. I'm too busy these days.